Return AndWhichConstraint<ObjectAssertions, IEnumerable>
