+++ Src/FluentAssertions/Primitives/ObjectAssertions.cs
@@ -2,6 +2,7 @@
 using System.Collections;
 using System.Collections.Generic;
 using System.Reflection;
+using System.Threading;
 using FluentAssertions.Common;
 using FluentAssertions.Equivalency;
 using FluentAssertions.Execution;
@@ -254,7 +255,7 @@ namespace FluentAssertions.Primitives
         }
 
         /// <summary>
-        /// Asserts that an object is enumerable by a foreach loop.g
+        /// Asserts that an object is enumerable using a foreach loop.
         /// </summary>
         /// <param name="because">
         /// A formatted phrase explaining why the assertion should be satisfied. If the phrase does not
@@ -263,11 +264,11 @@ namespace FluentAssertions.Primitives
         /// <param name="becauseArgs">
         /// Zero or more values to use for filling in any <see cref="string.Format(string,object[])" /> compatible placeholders.
         /// </param>
-        public AndConstraint<ObjectAssertions> BeEnumerable(string because = "", params object[] becauseArgs)
+        public AndWhichConstraint<ObjectAssertions, IEnumerable> BeEnumerable(string because = "", params object[] becauseArgs)
         {
             AssertIsEnumerable(Subject, because, becauseArgs);
 
-            return new AndConstraint<ObjectAssertions>(this);
+            return new AndWhichConstraint<ObjectAssertions, IEnumerable>(this, new ObjectEnumerable(Subject));
         }
 
         public AndConstraint<ObjectAssertions> BeEnumerable(IEnumerable expected, Func<object, object, bool> equalityComparison, string because = "", params object[] becauseArgs)
@@ -330,6 +331,61 @@ namespace FluentAssertions.Primitives
                 .WithExpectation("Expected {context:collection} to be equal to {0}{reason}, ", expectation)
                 .Given(() => subject)
                 .AssertForEachEnumerablesHaveSameItems(expectation, (a, e) => a.IndexOfFirstDifferenceWith(enumerableType, e, equalityComparison));
+
+        internal class ObjectEnumerable : IEnumerable
+        {
+            readonly object items;
+            MethodInfo methodGetEnumerator;
+
+            public ObjectEnumerable(object items)
+            {
+                this.items = items;
+            }
+
+            public IEnumerator GetEnumerator() => new Enumerator(this);
+
+            class Enumerator : IEnumerator, IDisposable
+            {
+                readonly object enumerator;
+                readonly PropertyInfo propertyCurrent;
+                readonly MethodInfo methodMoveNext;
+                MethodInfo methodReset;
+                MethodInfo methodDispose;
+                bool methodResetInitialized;
+                bool methodDisposeInitialized;
+                object methodResetSync;
+                object methodDisposeSync;
+
+                public Enumerator(ObjectEnumerable enumerable)
+                {
+                    LazyInitializer.EnsureInitialized(ref enumerable.methodGetEnumerator,
+                        () => enumerable.items.GetType().GetPublicOrExplicitParameterlessMethod("GetEnumerator"));
+                    enumerator = enumerable.methodGetEnumerator.Invoke(enumerable.items, new object[0]);
+
+                    var enumeratorType = enumerator.GetType();
+                    propertyCurrent = enumeratorType.GetPublicOrExplicitProperty("Current");
+                    methodMoveNext = enumeratorType.GetPublicOrExplicitParameterlessMethod("MoveNext");
+                }
+
+                public object Current => propertyCurrent.GetValue(enumerator);
+
+                public bool MoveNext() => (bool)methodMoveNext.Invoke(enumerator, new object[0]);
+
+                public void Reset()
+                {
+                    LazyInitializer.EnsureInitialized(ref methodReset, ref methodResetInitialized, ref methodResetSync,
+                        () => enumerator.GetType().GetPublicOrExplicitParameterlessMethod("Reset"));
+                    methodReset?.Invoke(enumerator, new object[0]);
+                }
+
+                public void Dispose()
+                {
+                    LazyInitializer.EnsureInitialized(ref methodDispose, ref methodDisposeInitialized, ref methodDisposeSync,
+                        () => enumerator.GetType().GetPublicOrExplicitParameterlessMethod("Dispose"));
+                    methodDispose?.Invoke(enumerator, new object[0]);
+                }
+            }
+        }
 
         /// <summary>
         /// Returns the type of the subject the assertion applies on.
